pyproject.toml
[project]
name = "Karakeep Auto Adder"
version = "0.1.0"
description = "Server side of a Telegram Bot to automatically add links sent to the bot to a local Karakeep instance."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
]
